# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsultMessage": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "sender_physician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Physician Id"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "thread_id",
          "body",
          "created_at"
        ],
        "title": "ConsultMessage",
        "type": "object"
      },
      "ConsultThread": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initiator_physician_id": {
            "title": "Initiator Physician Id",
            "type": "string"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsultMessage"
              },
              {
                "type": "null"
              }
            ]
          },
          "recipient_physician_id": {
            "title": "Recipient Physician Id",
            "type": "string"
          },
          "referral_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Id"
          },
          "unread_count": {
            "default": 0,
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "initiator_physician_id",
          "recipient_physician_id",
          "created_at"
        ],
        "title": "ConsultThread",
        "type": "object"
      },
...
      "GeoPoint": {
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "GeoPoint",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InsurancePlan": {
        "properties": {
          "payer": {
            "title": "Payer",
            "type": "string"
          },
          "plan_name": {
            "title": "Plan Name",
            "type": "string"
          }
        },
        "required": [
          "payer",
          "plan_name"
        ],
        "title": "InsurancePlan",
        "type": "object"
      },
...
      "Patient": {
        "description": "Synthetic patient. Never put real PHI here.",
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "insurance": {
            "$ref": "#/components/schemas/InsurancePlan"
          },
          "location": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "zip_code": {
            "title": "Zip Code",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "age",
          "zip_code",
          "location",
          "insurance"
        ],
        "title": "Patient",
        "type": "object"
      },
...
      "Physician": {
        "description": "A referring physician (synthetic). Stands in for the authenticated user until auth lands.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "practice_name": {
            "title": "Practice Name",
            "type": "string"
          },
          "specialty": {
            "title": "Specialty",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "specialty",
          "practice_name"
        ],
        "title": "Physician",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "sender_physician_id": {
            "title": "Sender Physician Id",
            "type": "string"
          }
        },
        "required": [
          "sender_physician_id",
          "body"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
      "Specialist": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "practice_name": {
            "title": "Practice Name",
            "type": "string"
          },
          "specialty": {
            "title": "Specialty",
            "type": "string"
          },
          "subspecialties": {
            "items": {
              "type": "string"
            },
            "title": "Subspecialties",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "specialty",
          "subspecialties",
          "practice_name",
          "address",
          "location"
        ],
        "title": "Specialist",
        "type": "object"
      },
      "StartConsultRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "from_physician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Physician Id"
          },
          "referral_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Id"
          },
          "to_physician_id": {
            "title": "To Physician Id",
            "type": "string"
          }
        },
        "required": [
          "to_physician_id",
          "body"
        ],
        "title": "StartConsultRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Dev mode only (AUTH_MODE=dev): act as this physician id, e.g. doc_001",
        "in": "header",
        "name": "X-Dev-Physician",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "Auth0 access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Physician-first referral matching. Nothing is sent or booked without physician approval.",
    "title": "ECHO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/consults": {
      "get": {
        "description": "The signed-in physician's threads, each with its last message and unread count.\n`physician_id` is optional; if sent it must be the signed-in physician.",
        "operationId": "list_consults_consults_get",
        "parameters": [
          {
            "in": "query",
            "name": "physician_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Physician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsultThread"
                  },
                  "title": "Response List Consults Consults Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Consults",
        "tags": [
          "consult"
        ]
      },
      "post": {
        "operationId": "start_consult_consults_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartConsultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultThread"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Consult",
        "tags": [
          "consult"
        ]
      }
    },
    "/consults/{thread_id}/messages": {
      "get": {
        "description": "Full history; marks the other participant's messages as read by the signed-in physician.\n`reader_physician_id` is optional; if sent it must be the signed-in physician.",
        "operationId": "get_consult_messages_consults__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reader_physician_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reader Physician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsultMessage"
                  },
                  "title": "Response Get Consult Messages Consults  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Consult Messages",
        "tags": [
          "consult"
        ]
      },
      "post": {
        "operationId": "send_consult_message_consults__thread_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultMessage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Send Consult Message",
        "tags": [
          "consult"
        ]
      }
    },
...
    "/patients": {
      "get": {
        "operationId": "list_patients_patients_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  },
                  "title": "Response List Patients Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Patients",
        "tags": [
          "directory"
        ]
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Patient",
        "tags": [
          "directory"
        ]
      }
    },
    "/physicians": {
      "get": {
        "operationId": "list_physicians_physicians_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Physician"
                  },
                  "title": "Response List Physicians Physicians Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Physicians",
        "tags": [
          "directory"
        ]
      }
    },
...
    "/specialists": {
      "get": {
        "operationId": "list_specialists_specialists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Specialist"
                  },
                  "title": "Response List Specialists Specialists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Specialists",
        "tags": [
          "directory"
        ]
      }
    },
    "/specialists/{specialist_id}": {
      "get": {
        "operationId": "get_specialist_specialists__specialist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "specialist_id",
            "required": true,
            "schema": {
              "title": "Specialist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Specialist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Specialist",
        "tags": [
          "directory"
        ]
      }
    }
  }
}